refac: 하트 부족 에러 핸들링 하트 도메인에서 처리
